+++ cobras/runner/health.py
@@ -12,21 +12,32 @@ from cobras.client.credentials import getDefaultRoleForApp, getDefaultSecretForA
 from cobras.client.health_check import (
     getDefaultHealthCheckChannel,
     getDefaultHealthCheckHttpUrl,
-    getDefaultHealthCheckUrl,
     healthCheck,
 )
+from cobras.common.apps_config import HEALTH_APPKEY, getDefaultEndpoint, makeUrl
 
 
-@click.option('--url', default=getDefaultHealthCheckUrl())
+@click.command()
+@click.option('--endpoint', default=getDefaultEndpoint())
+@click.option('--appkey', default=HEALTH_APPKEY)
 @click.option('--http_url', default=getDefaultHealthCheckHttpUrl)
 @click.option('--http', is_flag=True)
 @click.option('--http_extra_check', is_flag=True)
-@click.option('--role', default=getDefaultRoleForApp('health'))
-@click.option('--secret', default=getDefaultSecretForApp('health'))
+@click.option('--rolename', default=getDefaultRoleForApp('health'))
+@click.option('--rolesecret', default=getDefaultSecretForApp('health'))
 @click.option('--channel', default=getDefaultHealthCheckChannel())
 @click.option('--retry', is_flag=True)
-@click.command()
-def health(url, http_url, http, http_extra_check, role, secret, channel, retry):
+def health(
+    endpoint,
+    appkey,
+    http_url,
+    http,
+    http_extra_check,
+    rolename,
+    rolesecret,
+    channel,
+    retry,
+):
     '''Health check
 
     \b
@@ -35,6 +46,7 @@ def health(url, http_url, http, http_extra_check, role, secret, channel, retry):
     cobra health --http --http_url 'http://127.0.0.1:8765/health/'
     \b
     '''
+    url = makeUrl(endpoint, appkey)
 
     if http:
         print('url:', http_url)
@@ -43,7 +55,7 @@ def health(url, http_url, http, http_extra_check, role, secret, channel, retry):
             print(html.decode('utf8'), end='')
     else:
         try:
-            healthCheck(url, role, secret, channel, retry, http_extra_check)
+            healthCheck(url, rolename, rolesecret, channel, retry, http_extra_check)
         except Exception as e:
             click.secho(f'System is unhealthy !!: {e}', fg='red')
             sys.exit(1)